oxygen: initial bring up for twrp and LinageOS
