+++ Android.mk
@@ -14,9 +14,9 @@
 # limitations under the License.
 #
 
-LOCAL_PATH := $(call my-dir)
+ifneq ($(filter oxygen,$(TARGET_DEVICE)),)
 
-ifneq ($(filter mido,$(TARGET_DEVICE)),)
+LOCAL_PATH := $(call my-dir)
 
 include $(call all-makefiles-under,$(LOCAL_PATH))
