+++ audio/audio_platform_info.xml
@@ -53,20 +53,57 @@
         <param key="spkr_2_tz_name" value=""/>
     </config_params>
     <acdb_ids>
-        <device name="SND_DEVICE_OUT_SPEAKER_PROTECTED" acdb_id="136"/>
-        <device name="SND_DEVICE_OUT_VOICE_SPEAKER_PROTECTED" acdb_id="136"/>
+       <device name="SND_DEVICE_OUT_HANDSET" acdb_id="18"/>
+        <device name="SND_DEVICE_OUT_SPEAKER" acdb_id="34"/>
+        <device name="SND_DEVICE_OUT_HEADPHONES" acdb_id="52"/>
+        <device name="SND_DEVICE_OUT_SPEAKER_AND_HEADPHONES" acdb_id="124"/>
+        <device name="SND_DEVICE_OUT_VOICE_HANDSET" acdb_id="18"/>
+        <device name="SND_DEVICE_OUT_VOICE_SPEAKER" acdb_id="34"/>
+        <device name="SND_DEVICE_OUT_VOICE_HEADPHONES" acdb_id="52"/>
+        <device name="SND_DEVICE_OUT_BT_SCO" acdb_id="66"/>
+        <device name="SND_DEVICE_OUT_BT_SCO_WB" acdb_id="578"/>
+        <device name="SND_DEVICE_OUT_VOICE_TX" acdb_id="52"/>
+        <device name="SND_DEVICE_OUT_TRANSMISSION_FM" acdb_id="0"/>
+        <device name="SND_DEVICE_OUT_SPEAKER_PROTECTED" acdb_id="124"/>
+        <device name="SND_DEVICE_IN_VOICE_REC_QMIC_FLUENCE" acdb_id="131"/>
+        <device name="SND_DEVICE_IN_VOICE_REC_TMIC" acdb_id="131"/>
+        <device name="SND_DEVICE_IN_VOICE_REC_DMIC_FLUENCE" acdb_id="132"/>
+        <device name="SND_DEVICE_OUT_VOICE_SPEAKER_PROTECTED" acdb_id="546"/>
+        <device name="SND_DEVICE_IN_HANDSET_MIC" acdb_id="19"/>
+        <device name="SND_DEVICE_IN_HANDSET_DMIC" acdb_id="41"/>
+        <device name="SND_DEVICE_IN_SPEAKER_MIC" acdb_id="35"/>
+        <device name="SND_DEVICE_IN_SPEAKER_DMIC" acdb_id="37"/>
+        <device name="SND_DEVICE_IN_HEADSET_MIC" acdb_id="51"/>
+        <device name="SND_DEVICE_IN_VOICE_SPEAKER_MIC" acdb_id="291"/>
+        <device name="SND_DEVICE_IN_VOICE_HEADSET_MIC" acdb_id="307"/>
+        <device name="SND_DEVICE_IN_BT_SCO_MIC" acdb_id="323"/>
+        <device name="SND_DEVICE_IN_BT_SCO_MIC_NREC" acdb_id="835"/>
+        <device name="SND_DEVICE_IN_BT_SCO_MIC_WB" acdb_id="1347"/>
+        <device name="SND_DEVICE_IN_BT_SCO_MIC_WB_NREC" acdb_id="1859"/>
+        <device name="SND_DEVICE_IN_CAMCORDER_MIC" acdb_id="4"/>
+        <device name="SND_DEVICE_IN_VOICE_DMIC" acdb_id="277"/>
+        <device name="SND_DEVICE_IN_VOICE_SPEAKER_DMIC" acdb_id="293"/>
+        <device name="SND_DEVICE_IN_VOICE_RX" acdb_id="819"/>
+        <device name="SND_DEVICE_IN_VOICE_REC_MIC" acdb_id="4"/>
+        <device name="SND_DEVICE_IN_CAPTURE_FM" acdb_id="0"/>
+        <device name="SND_DEVICE_IN_CAPTURE_VI_FEEDBACK" acdb_id="102"/>
+        <device name="SND_DEVICE_IN_MAIN_MIC" acdb_id="4"/>
+        <device name="SND_DEVICE_IN_TOP_MIC" acdb_id="11"/>
+        <device name="SND_DEVICE_IN_FRONT_MIC" acdb_id="11"/>
+        <device name="SND_DEVICE_IN_HANDSET_DMIC_MUSIC" acdb_id="21"/>
+        <device name="SND_DEVICE_IN_HANDSET_DMIC_VOICE" acdb_id="533"/>
+        <device name="SND_DEVICE_IN_HANDSET_DMIC_INTERVIEW" acdb_id="1045"/>
+        <device name="SND_DEVICE_IN_HANDSET_DMIC_HD" acdb_id="2581"/>
+        <device name="SND_DEVICE_IN_HANDSET_STEREO_DMIC" acdb_id="2069"/>
+        <device name="SND_DEVICE_IN_VOICE_SPEAKER_TMIC" acdb_id="295"/>
+        <device name="SND_DEVICE_IN_VOICE_SPEAKER_TMIC_CONF" acdb_id="1319"/>
+        <device name="SND_DEVICE_IN_CAMCORDER_DMIC" acdb_id="85"/>
+        <device name="SND_DEVICE_IN_CAMCORDER_TMIC" acdb_id="87"/>
+        <device name="SND_DEVICE_IN_CAMCORDER_TMIC_FAR_END" acdb_id="599"/>
+        <device name="SND_DEVICE_IN_CAMCORDER_TMIC_NEAR_END" acdb_id="1111"/>
         <device name="SND_DEVICE_OUT_SPEAKER_PROTECTED_VBAT" acdb_id="136"/>
         <device name="SND_DEVICE_OUT_VOICE_SPEAKER_PROTECTED_VBAT" acdb_id="136"/>
         <device name="SND_DEVICE_IN_CAPTURE_VI_FEEDBACK" acdb_id="137"/>
-
-	<device name="SND_DEVICE_IN_SPEAKER_DMIC_AEC_BROADSIDE" acdb_id="121"/>
-        <device name="SND_DEVICE_IN_SPEAKER_DMIC_AEC_NS_BROADSIDE" acdb_id="105"/>
-        <device name="SND_DEVICE_IN_SPEAKER_DMIC_BROADSIDE" acdb_id="119"/>
-        <device name="SND_DEVICE_IN_SPEAKER_DMIC_NS_BROADSIDE" acdb_id="120"/>
-        <device name="SND_DEVICE_IN_VOICE_FLUENCE_DMIC_AANC" acdb_id="12"/>
-        <device name="SND_DEVICE_OUT_SPEAKER_REVERSE" acdb_id="135"/>
-        <device name="SND_DEVICE_OUT_SPEAKER_WSA" acdb_id="14"/>	
-
     </acdb_ids>
 </audio_platform_info>
