+++ configs/media_codecs_performance.xml
@@ -78,8 +78,8 @@
             <Limit name="measured-frame-rate-352x288" range="434-434" />
             <Limit name="measured-frame-rate-720x480" range="338-338" />
             <Limit name="measured-frame-rate-1280x720" range="241-241" />
-            <Limit name="measured-frame-rate-1920x1088" range="50-60" />
-            <Limit name="measured-frame-rate-3840x2160" range="10-13" />
+            <Limit name="measured-frame-rate-1920x1088" range="180-180" />
+            <Limit name="measured-frame-rate-3840x2160" range="44-44" />
         </MediaCodec>
         <MediaCodec name="OMX.qcom.video.decoder.h263" type="video/3gpp" update="true">
             <Limit name="measured-frame-rate-176x144" range="400-400" />
