+++ configs/thermal-engine.conf
@@ -1,19 +1,2 @@
 # File empty by default.
 # Replace contents of this file with custom configuration.
-[MODEM_PA_MITIGATION]
-#algo_type monitor
-sampling 10000
-sensor case_therm
-thresholds 48000 58000
-thresholds_clr 45000 54000
-actions modem modem
-action_info 2 3
-
-[MODEM_PROC_TEMP_MITIGATION]
-#algo_type monitor
-sampling 10000
-sensor case_therm
-thresholds 45000 58000
-thresholds_clr 43000 54000
-actions modem_proc modem_proc
-action_info 1 3
