+++ init/Android.mk
@@ -19,8 +19,8 @@ LOCAL_PATH := $(call my-dir)
 include $(CLEAR_VARS)
 
 LOCAL_C_INCLUDES := system/core/init
-LOCAL_MODULE := libinit_mido
+LOCAL_MODULE := libinit_oxygen
 LOCAL_MODULE_TAGS := optional
-LOCAL_SRC_FILES := init_mido.cpp
+LOCAL_SRC_FILES := init_oxygen.cpp
 
 include $(BUILD_STATIC_LIBRARY)
