+++ lineage.dependencies
@@ -3,8 +3,4 @@
    "repository": "android_device_qcom_common",
    "target_path": "device/qcom/common"
  },
- {
-   "repository": "android_kernel_xiaomi_msm8953",
-   "target_path": "kernel/xiaomi/msm8953"
- }
 ]
