+++ rootdir/adb_keys
@@ -1,0 +1,1 @@
+QAAAAPnT61C3gTOPjhpzTHBR1FimcbtCVsPjzvNBJeHteNNHHRnsOaYdwQIgf+1qj9sdoY/QkXMksHira0Cid4bXmuwxU1MdxQy3RU+liSCS8iquUFSBYNLfYJHQ4cJvvYv+vJamtY1nI6lPfYzqWxuFcOku0+SqMHVK+sD3TNx7dm7nuJx9qbonxsd6xk5o62CPoHubSeVx/KN3KuSGsEwB4sGRcXZ0Iz/OPEAwZ8YzMqMqoliIbxqChVvCu2fy/GBUmXxJ9miQNRbviZYRzvWOCoB8al2grDD+1tFyhAZ7tyEfHQLWLiFayx9Ap1c85sDRyLPv9XIDKOBc4lVGd+8TJzCaQUS/y/ik3q88B+ZrcPw4g/yl6BafDPIIdtm5l+xXQUHh2eLm1+eQVrlvyqWDNRyBFY8LdxT1Eyn+ICCG1sSimRHZsSEntKn1Se7DvqrzrvZDVuVYNJ0//RwllR91U7Rvs1WvAhFg49q8+N7jXRqCxthoWcvTUGZXZoGvs8qcCVFGr2xxFwt+IJgpJsdzOvGEQ0HNwugbkGJsdbzCEy0nFkylV8PrqGxSB+ci7hlJgdB4DE/gruebgpmJ7mFyOOvWssSeexEzduJoqpxZ1bvx1G/jVQ/DknNo774SK0/To6F/1IbVe7xTm/X510u62CNKm7GdxaZ7xYwe84uJ1zbU6FxpqgEAAQA= xuefer@home
