+++ rootdir/fstab.qcom
@@ -6,16 +6,16 @@
 /dev/block/bootdevice/by-name/boot		/boot			emmc	defaults							defaults
 /dev/block/bootdevice/by-name/recovery		/recovery		emmc	defaults							defaults
 
-/dev/block/bootdevice/by-name/system		/system			ext4	ro								wait
-/dev/block/bootdevice/by-name/userdata		/data			f2fs	nosuid,nodev,noatime,inline_xattr,data_flush			wait,check,encryptable=footer,formattable,length=-16384
-/dev/block/bootdevice/by-name/userdata		/data			ext4	nosuid,nodev,noatime,noauto_da_alloc				wait,check,encryptable=footer,formattable,length=-16384
-/dev/block/bootdevice/by-name/cache		/cache			f2fs	nosuid,nodev,noatime,inline_xattr,flush_merge,data_flush	wait,formattable,check
+/dev/block/bootdevice/by-name/system		/system			ext4	ro,barrier=1,discard						wait
+#/dev/block/bootdevice/by-name/userdata		/data			f2fs	nosuid,nodev,noatime,inline_xattr,data_flush			wait,check,forceencrypt=footer,formattable,length=-16384
+/dev/block/bootdevice/by-name/userdata		/data			ext4	nosuid,nodev,barrier=1,noauto_da_alloc,noatime,discard			wait,check,forceencrypt=footer,formattable,length=-16384
+#/dev/block/bootdevice/by-name/cache		/cache			f2fs	nosuid,nodev,noatime,inline_xattr,flush_merge,data_flush	wait,formattable,check
 /dev/block/bootdevice/by-name/cache		/cache			ext4	nosuid,nodev,noatime						wait,formattable,check
-/dev/block/bootdevice/by-name/persist		/persist		ext4	nosuid,nodev							wait,check
+/dev/block/bootdevice/by-name/config		/frp			emmc	defaults							defaults
+/dev/block/bootdevice/by-name/misc		/misc			emmc	defaults							defaults
+/dev/block/bootdevice/by-name/persist		/persist		ext4	nosuid,nodev,barrier=1						wait,check,forceformat
 /dev/block/bootdevice/by-name/dsp		/dsp			ext4	ro,nosuid,nodev							wait
 /dev/block/bootdevice/by-name/modem		/firmware		vfat	ro,context=u:object_r:firmware_file:s0,shortname=lower,uid=1000,gid=1000,dmask=227,fmask=337 wait
-/dev/block/bootdevice/by-name/misc		/misc			emmc	defaults							defaults
-/dev/block/bootdevice/by-name/config		/frp			emmc	defaults							defaults
 
-/devices/soc/7864900.sdhci/mmc_host*		auto			auto	defaults							wait,voldmanaged=sdcard1:auto,noemulatedsd,encryptable=userdata
+/devices/soc/7864900.sdhci/mmc_host*		auto			auto	nosuid,nodev							wait,voldmanaged=sdcard1:auto,noemulatedsd,encryptable=userdata
 /devices/soc/7000000.ssusb/7000000.dwc3/xhci-hcd.0.auto*     auto	auto    defaults							wait,voldmanaged=usb:auto
