+++ rootdir/init.qcom.usb.sh
@@ -107,6 +107,7 @@ baseband=`getprop ro.baseband`
 
 echo 1  > /sys/class/android_usb/f_mass_storage/lun/nofua
 usb_config=`getprop persist.sys.usb.config`
+debuggable=`getprop ro.debuggable`
 case "$usb_config" in
     "" | "adb") #USB persist config not set, select default configuration
       case "$esoc_link" in
@@ -148,9 +149,13 @@ case "$usb_config" in
 			    esac
 		      ;;
 	              "msm8952" | "msm8953")
-		          setprop persist.sys.usb.config mtp,adb
+		          if [ -z "$debuggable" -o "$debuggable" = "1" ]; then
+                              setprop persist.sys.usb.config mtp,adb
+		          else
+                              setprop persist.sys.usb.config mtp
+		          fi
 		      ;;
-	              "msmcobalt")
+	              "msm8998")
 		          setprop persist.sys.usb.config diag,serial_cdev,rmnet_gsi,adb
 		      ;;
 	              *)
@@ -171,10 +176,16 @@ esac
 case "$target" in
     "msm8996")
         setprop sys.usb.controller "6a00000.dwc3"
+        setprop sys.usb.rndis.func.name "rndis_bam"
 	;;
-    "msmcobalt")
+    "msm8998")
         setprop sys.usb.controller "a800000.dwc3"
+        setprop sys.usb.rndis.func.name "gsi"
 	;;
+    "msmfalcon")
+        setprop sys.usb.controller "a800000.dwc3"
+        setprop sys.usb.rndis.func.name "rndis_bam"
+        ;;
     *)
 	;;
 esac
