+++ wifi/WCNSS_qcom_cfg.ini
@@ -398,7 +398,7 @@ gEnableMacAddrSpoof=1
 #Set RPS CPU mask
 rps_mask=0f
 
-#wifi ping test
+#Force split scan
 gEnableDynSplitScan=0
 
 END
